requirements.txt
python-binance
websocket-client
psycopg2
redis
flask
joblib
pandas
numpy
scikit-learn
tweepy
